Move `{Copy,Move}File` into dedicated package
This better belongs into its own package since it has not much to do with the `registry` pkg
