+++ pkg/nodeagent/registry/fake/fake_extractor.go
@@ -22,6 +22,7 @@ import (
 
 	"github.com/spf13/afero"
 
+	"github.com/gardener/gardener/pkg/nodeagent/files"
 	"github.com/gardener/gardener/pkg/nodeagent/registry"
 )
 
@@ -40,7 +41,7 @@ func NewExtractor(fakeFS afero.Afero, sourceDirectory string) registry.Extractor
 // CopyFromImage copies a file from a given image reference to the destination file.
 func (e *fakeRegistryExtractor) CopyFromImage(_ context.Context, _ string, filePathInImage string, destination string, permissions fs.FileMode) error {
 	source := path.Join(e.sourceDirectory, filePathInImage)
-	if err := registry.CopyFile(e.fakeFS, source, destination, permissions); err != nil {
+	if err := files.Copy(e.fakeFS, source, destination, permissions); err != nil {
 		return fmt.Errorf("error copying file %s to %s: %w", source, destination, err)
 	}
