+++ skaffold.yaml
@@ -1195,6 +1195,7 @@ build:
             - pkg/nodeagent/controller/token
             - pkg/nodeagent/dbus
             - pkg/nodeagent/features
+            - pkg/nodeagent/files
             - pkg/nodeagent/registry
             - pkg/resourcemanager/controller/garbagecollector/references
             - pkg/utils
